Reorganized files and deleted obsolete sl files
