+++ src/analysis/retail/monk/mistweaver/CombatLogParser.ts
@@ -1,12 +1,7 @@
 import {
   FaelineStomp,
-  FallenOrder,
-  FortifyingIngredients,
-  GroundingBreath,
-  ImbuedReflections,
   InvokersDelight,
   MysticTouch,
-  SinisterTeachings,
   TouchOfDeath,
   DampenHarm,
 } from 'analysis/retail/monk/shared';
@@ -30,16 +25,13 @@ import MasteryStats from './modules/features/MasteryStats';
 import MistweaverHealingEfficiencyDetails from './modules/features/MistweaverHealingEfficiencyDetails';
 import HealingEfficiencyTracker from './modules/features/MistweaverHealingEfficiencyTracker';
 import REMGraph from './modules/features/REMGraph';
-import JadeBond from './modules/shadowlands/conduits/JadeBond';
-import NourishingChi from './modules/shadowlands/conduits/NourishingChi';
-import RisingSunRevival from './modules/shadowlands/conduits/RisingSunRevival';
-import BonedustBrewAverageTargets from './modules/shadowlands/covenant/BonedustBrewAverageTargets';
-import FaelineStompHealing from './modules/shadowlands/covenant/FaelineStompHealing';
-import FallenOrderAverageHPOfTargetOnCast from './modules/shadowlands/covenant/FallenOrderAverageHPOfTargetOnCast';
-import FallenOrderCraneAverage from './modules/shadowlands/covenant/FallenOrderCraneAverage';
-import AncientTeachingsoftheMonastery from './modules/shadowlands/legendaries/AncientTeachingsoftheMonastery';
-import CloudedFocus from './modules/shadowlands/legendaries/CloudedFocus';
-import T28FourSet from './modules/shadowlands/tier/T28FourSet';
+import JadeBond from './modules/spells/JadeBond';
+import NourishingChi from './modules/spells/NourishingChi';
+import RisingSunRevival from './modules/spells/RisingSunRevival';
+import BonedustBrewAverageTargets from './modules/spells/BonedustBrewAverageTargets';
+import FaelineStompHealing from './modules/spells/FaelineStompHealing';
+import AncientTeachingsoftheMonastery from './modules/spells/AncientTeachingsoftheMonastery';
+import CloudedFocus from './modules/spells/CloudedFocus';
 import EnvelopingBreath from './modules/spells/EnvelopingBreath';
 import EnvelopingMists from './modules/spells/EnvelopingMists';
 import EssenceFont from './modules/spells/EssenceFont';
@@ -56,19 +48,18 @@ import SoothingMist from './modules/spells/SoothingMist';
 import SpinningCraneKick from './modules/spells/SpinningCraneKick';
 import ThunderFocusTea from './modules/spells/ThunderFocusTea';
 import Vivify from './modules/spells/Vivify';
-import AverageTimeBetweenRSKSs from './modules/talents/AverageTimeBetweenRSKs';
-import ChiBurst from './modules/talents/ChiBurst';
-import InvokeChiJi from './modules/talents/InvokeChiJi';
-import JadeSerpentStatue from './modules/talents/JadeSerpentStatue';
-import Lifecycles from './modules/talents/Lifecycles';
-import ManaTea from './modules/talents/ManaTea';
-import MistWrapEnvelopingBreath from './modules/talents/MistWrapEnvelopingBreath';
-import RefreshingJadeWind from './modules/talents/RefreshingJadeWind';
-import RenewingMistDuringManaTea from './modules/talents/RenewingMistDuringManaTea';
-import RisingMist from './modules/talents/RisingMist';
-import SpiritOfTheCrane from './modules/talents/SpiritOfTheCrane';
-import Upwelling from './modules/talents/Upwelling';
-import FallenOrderCraneCloneNormalizer from './normalizers/FallenOrderCraneCloneNormalizer';
+import AverageTimeBetweenRSKSs from './modules/spells/AverageTimeBetweenRSKs';
+import ChiBurst from './modules/spells/ChiBurst';
+import InvokeChiJi from './modules/spells/InvokeChiJi';
+import JadeSerpentStatue from './modules/spells/JadeSerpentStatue';
+import Lifecycles from './modules/spells/Lifecycles';
+import ManaTea from './modules/spells/ManaTea';
+import MistWrapEnvelopingBreath from './modules/spells/MistWrapEnvelopingBreath';
+import RefreshingJadeWind from './modules/spells/RefreshingJadeWind';
+import RenewingMistDuringManaTea from './modules/spells/RenewingMistDuringManaTea';
+import RisingMist from './modules/spells/RisingMist';
+import SpiritOfTheCrane from './modules/spells/SpiritOfTheCrane';
+import Upwelling from './modules/spells/Upwelling';
 import HotApplicationNormalizer from './normalizers/HotApplicationNormalizer';
 import HotRemovalNormalizer from './normalizers/HotRemovalNormalizer';
 
@@ -77,7 +68,6 @@ class CombatLogParser extends CoreCombatLogParser {
     // Normalizer
     hotApplicationNormalizer: HotApplicationNormalizer,
     hotRemovalNormalizer: HotRemovalNormalizer,
-    fallenOrderCraneCloneNormalizer: FallenOrderCraneCloneNormalizer,
 
     // Core
     lowHealthHealing: LowHealthHealing,
@@ -102,72 +92,54 @@ class CombatLogParser extends CoreCombatLogParser {
     averageTimeBetweenRSKSs: AverageTimeBetweenRSKSs,
     remGraph: REMGraph,
 
-    // Spells
-    essenceFont: EssenceFont,
-    thunderFocusTea: ThunderFocusTea,
-    envelopingMists: EnvelopingMists,
-    soothingMist: SoothingMist, // Removed as this needs to be reworked with updated Soothing Mist Spell in BfA
+    // Base Spells
     spinningCraneKick: SpinningCraneKick,
     vivify: Vivify,
-    renewingMist: RenewingMist,
-    lifeCocoon: LifeCocoon,
+
+    // Shared Talents
+    chiBurst: ChiBurst,
+    dampenHarm: DampenHarm,
     touchOfDeath: TouchOfDeath,
-    invokeYulon: InvokeYulon,
-    expelHarm: ExpelHarm,
-    envelopingBreath: EnvelopingBreath,
-    revival: Revival,
     risingSunKick: RisingSunKick,
 
+    // MW Talents
+    ancientTeachingsoftheMonastery: AncientTeachingsoftheMonastery,
+    bonedustBrewAverageTargets: BonedustBrewAverageTargets,
+    cloudedFocus: CloudedFocus,
+    envelopingBreath: EnvelopingBreath,
+    envelopingMists: EnvelopingMists,
+    essenceFont: EssenceFont,
     essenceFontUniqueTargets: EssenceFontUniqueTargets,
     essenceFontTargetsHit: EssenceFontTargetsHit,
+    expelHarm: ExpelHarm,
     EssenceFontCancelled: EssenceFontCancelled,
-
-    // Talents
-    chiBurst: ChiBurst,
-    manaTea: ManaTea,
-    refreshingJadeWind: RefreshingJadeWind,
-    lifecycles: Lifecycles,
-    spiritOfTheCrane: SpiritOfTheCrane,
-    risingMist: RisingMist,
-    jadeSerpentStatue: JadeSerpentStatue,
-    renewingMistDuringManaTea: RenewingMistDuringManaTea,
-    upwelling: Upwelling,
+    faelineStomp: FaelineStomp,
+    faelineStompHealing: FaelineStompHealing,
+    invokersDelight: InvokersDelight,
     invokeChiJi: InvokeChiJi,
+    invokeYulon: InvokeYulon,
+    jadeSerpentStatue: JadeSerpentStatue,
+    jadeBond: JadeBond,
+    lifecycles: Lifecycles,
+    lifeCocoon: LifeCocoon,
     mistWrapEnvelopingBreath: MistWrapEnvelopingBreath,
-    dampenHarm: DampenHarm,
+    manaTea: ManaTea,
+    nourishingCh: NourishingChi,
+    refreshingJadeWind: RefreshingJadeWind,
+    renewingMist: RenewingMist,
+    renewingMistDuringManaTea: RenewingMistDuringManaTea,
+    revival: Revival,
+    risingMist: RisingMist,
+    risingSunRevival: RisingSunRevival,
+    spiritOfTheCrane: SpiritOfTheCrane,
+    soothingMist: SoothingMist,
+    thunderFocusTea: ThunderFocusTea,
+    upwelling: Upwelling,
 
     // Mana Tab
     manaTracker: ManaTracker,
     hpmDetails: MistweaverHealingEfficiencyDetails,
     hpmTracker: HealingEfficiencyTracker,
-
-    // Covenants
-    fallenOrder: FallenOrder,
-    fallenOrderCraneAverage: FallenOrderCraneAverage,
-    fallenOrderAverageHPOfTargetOnCast: FallenOrderAverageHPOfTargetOnCast,
-    faelineStomp: FaelineStomp,
-    faelineStompHealing: FaelineStompHealing,
-    bonedustBrewAverageTargets: BonedustBrewAverageTargets,
-
-    // Conduits
-    // Endurance
-    groundingBreath: GroundingBreath,
-    fortifyingIngredients: FortifyingIngredients,
-
-    // Potency
-    jadeBond: JadeBond,
-    nourishingChi: NourishingChi,
-    risingSunRevival: RisingSunRevival,
-    imbuedReflections: ImbuedReflections,
-
-    // Legendaries
-    ancientTeachingsoftheMonastery: AncientTeachingsoftheMonastery,
-    cloudedFocus: CloudedFocus,
-    invokersDelight: InvokersDelight,
-    sinisterTeachings: SinisterTeachings,
-
-    // Tier set
-    t28FourSet: T28FourSet,
   };
 }
