+++ src/analysis/retail/monk/mistweaver/modules/features/Checklist/Module.tsx
@@ -11,13 +11,13 @@ import SoothingMist from '../../spells/SoothingMist';
 import SpinningCraneKick from '../../spells/SpinningCraneKick';
 import ThunderFocusTea from '../../spells/ThunderFocusTea';
 import Vivify from '../../spells/Vivify';
-import ChiBurst from '../../talents/ChiBurst';
-import JadeSerpentStatue from '../../talents/JadeSerpentStatue';
-import Lifecycles from '../../talents/Lifecycles';
-import ManaTea from '../../talents/ManaTea';
-import RefreshingJadeWind from '../../talents/RefreshingJadeWind';
-import RenewingMistDuringManaTea from '../../talents/RenewingMistDuringManaTea';
-import SpiritOfTheCrane from '../../talents/SpiritOfTheCrane';
+import ChiBurst from '../../spells/ChiBurst';
+import JadeSerpentStatue from '../../spells/JadeSerpentStatue';
+import Lifecycles from '../../spells/Lifecycles';
+import ManaTea from '../../spells/ManaTea';
+import RefreshingJadeWind from '../../spells/RefreshingJadeWind';
+import RenewingMistDuringManaTea from '../../spells/RenewingMistDuringManaTea';
+import SpiritOfTheCrane from '../../spells/SpiritOfTheCrane';
 import AlwaysBeCasting from '../AlwaysBeCasting';
 import Component from './Component';
